+++ apps/api-harmonization/src/modules/page/page.mapper.ts
@@ -139,6 +139,7 @@ export const mapInit = (
     header: CMS.Model.Header.Header,
     footer: CMS.Model.Footer.Footer,
     labels: CMS.Model.AppConfig.Labels,
+    themes: CMS.Model.AppConfig.Themes,
     roles: Auth.Constants.Roles[],
 ): Init => {
     return {
@@ -174,5 +175,6 @@ export const mapInit = (
             },
         },
         labels,
+        themes,
     };
 };