+++ packages/integrations/strapi-cms/src/modules/cms/cms.service.ts
@@ -95,7 +95,7 @@ export class CmsService implements CMS.Service {
                 locale: options.locale,
             });
 
-            return forkJoin([appConfig]).pipe(map(([appConfig]) => mapAppConfig(appConfig.data)));
+            return forkJoin([appConfig]).pipe(map(([appConfig]) => mapAppConfig(appConfig.data, this.baseUrl)));
         });
     }
 
@@ -128,7 +128,7 @@ export class CmsService implements CMS.Service {
                     });
 
                     if (!page) {
-                        throw new NotFoundException();
+                        return undefined;
                     }
 
                     return mapPage(page);